refactor(RichTextEditor): shared useAnchoredFloating overlay hook
