+++ packages/design-system/src/components/RichTextEditor/RichTextLinkEditor.tsx
@@ -4,15 +4,15 @@
 // Floating UI virtual element (the same portal+virtual-anchor pattern as
 // LiquidEditor's AutocompleteMenu, so it escapes the editor's overflow and any
 // Drawer/Modal ancestor). Enter applies, Esc / click-outside cancels.
-import { useCallback, useEffect, useMemo, useRef, useState } from 'react';
+import { useCallback, useEffect, useRef, useState } from 'react';
 import { createPortal } from 'react-dom';
-import { useFloating, autoUpdate, flip, shift, offset } from '@floating-ui/react-dom';
 import { Button } from '../Button';
 import { Input } from '../Input';
 import { Stack } from '../Stack';
 import { Cluster } from '../Cluster';
 import { useTranslation } from '../../i18n';
 import type { Rect } from './selection';
+import { useAnchoredFloating } from './useAnchoredFloating';
 import styles from './RichTextEditor.module.scss';
 
 export interface RichTextLinkEditorProps {
@@ -56,35 +56,7 @@ export function RichTextLinkEditor({
   const inputRef = useRef<HTMLInputElement | null>(null);
   const bubbleRef = useRef<HTMLDivElement | null>(null);
 
-  // Floating UI virtual element — only `getBoundingClientRect` is required. Read the
-  // LIVE rect (via getAnchorRect) on every reposition so the bubble tracks the
-  // selection line on scroll; fall back to the static `anchorRect`.
-  const virtualRef = useMemo(
-    () => ({
-      getBoundingClientRect: () => {
-        const r = getAnchorRect?.() ?? anchorRect;
-        return {
-          x: r.left,
-          y: r.top,
-          top: r.top,
-          left: r.left,
-          right: r.left + r.width,
-          bottom: r.top + r.height,
-          width: r.width,
-          height: r.height,
-        };
-      },
-    }),
-    [anchorRect, getAnchorRect],
-  );
-
-  const { refs, floatingStyles } = useFloating({
-    placement: 'bottom-start',
-    strategy: 'fixed',
-    whileElementsMounted: autoUpdate,
-    middleware: [offset(6), flip(), shift({ padding: 4 })],
-    elements: { reference: virtualRef },
-  });
+  const { refs, floatingStyles } = useAnchoredFloating(anchorRect, getAnchorRect, { offset: 6 });
 
   // Focus the URL field on open; select its contents when editing so a re-type
   // replaces the existing href.
